+++ src/Molecules/Input/Select/Select.stories.tsx
@@ -209,16 +209,18 @@ export default {
 } as Meta;
 
 export const defaultStory = () => (
-  <Input.Select
-    id="input-1"
-    options={[
-      { value: 1, label: '1' },
-      { value: 2, label: '2' },
-      { value: 3, label: '3' },
-    ]}
-    label="Label"
-    placeholder="Placeholder"
-  />
+  <div style={{ padding: '10px', height: '200px' }}>
+    <Input.Select
+      id="input-1"
+      options={[
+        { value: 1, label: '1' },
+        { value: 2, label: '2' },
+        { value: 3, label: '3' },
+      ]}
+      label="Label"
+      placeholder="Placeholder"
+    />
+  </div>
 );
 
 export const withLabelTooltip = () => (
@@ -660,6 +662,55 @@ export const multiselect = () =>
       />
     );
   });
+
+const MultiselectColumnsTemplate: Story<{ itemsPerColumn: number; columnWidth: string }> = (
+  args,
+) => {
+  const MultiselectColumnsExample = () => {
+    const [value, setValue] = useState([]);
+
+    // This component you need to redefine for your particular case
+    // Consider translations and a11y!
+    const CustomSelectedValue = () => {
+      const [machineState] = useSelectMachineFromContext();
+      const selectedCount = machineState.context.selectedItems.length;
+      return (
+        <FlexedBox px={2}>
+          <Typography
+            type="secondary"
+            color={selectedCount === 0 ? (t) => t.color.placeholderText : undefined}
+          >
+            {selectedCount === 0 ? machineState.context.placeholder : `${selectedCount} selected`}
+          </Typography>
+        </FlexedBox>
+      );
+    };
+
+    return (
+      <Input.Select
+        id="custom-renderers-select"
+        options={accountOptions}
+        value={value}
+        // @ts-ignore
+        onChange={setValue}
+        components={{ SelectedValue: CustomSelectedValue }}
+        multiselect
+        label="User account"
+        placeholder="Select account"
+        itemsPerColumn={args.itemsPerColumn}
+        columnWidth={args.columnWidth}
+      />
+    );
+  };
+
+  return <MultiselectColumnsExample />;
+};
+
+export const MultiselectColumns = MultiselectColumnsTemplate.bind({});
+MultiselectColumns.args = {
+  itemsPerColumn: 4,
+  columnWidth: '200px',
+};
 
 export const multiselectActions = () =>
   createElement(() => {