+++ src/Molecules/Input/Select/lib/SingleSelectList/SingleSelectList.tsx
@@ -149,7 +149,7 @@ const StyledOption = styled(Typography)<Partial<OptionProps>>`
   ${(p) =>
     p.disabled
       ? `
-        color: ${p.theme.color.disabledText}
+        color: ${p.theme.color.disabledText};
         pointer-events: none;
       `
       : ``}